+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### UNRELEASED
+### 0.20.0 - 2020-04-12
 
  - Add docs for background tasks with lock args and redis